backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.133.0",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
    "langchain-ollama>=0.2.3",
    "langchain-groq>=1.1.2",
    "email-validator>=2.2.0",
    "google-api-python-client>=2.190.0",
    "google-auth-oauthlib>=1.3.0",
    "slack-sdk>=3.40.1",
    "anyio>=4.12.1",
    "sendgrid>=6.12.5",
    "pydantic-settings>=2.13.1",
]
